autem coeruleo
For Assignment #7 your program will enter a phone number as a string that contains letters (e.g. 1-800-Flowers) and returns a phone number of all digits (e.g. 1-800-3569377). 

Write a method with this header: public static int  getNumber (char uppercaseLetter)
Your method will receive a upper-case letter and return the corresponding number. 
Your challenge is to assure an uppercase character is passed to the method, and to properly handle characters that are not A-Z.